refactor: Refactor BorrowTransaction to use navigation properties and change the return type of BorrowBook.
